requirements.txt
asgiref==3.2.7
astroid==2.4.1
Django==3.0.6
django-appconf==1.0.4
django-compressor==2.4
django-fontawesome==1.0
django-libsass==0.8
django-sass-processor==0.8
django-scss==0.6.1
django-taggit==1.2.0
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
libsass==0.20.0
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.2
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
